requirements.txt
streamlit
anthropic
newspaper3k
PyPDF2
lxml_html_clean
